requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
apscheduler>=3.10
python-dotenv>=1.0
pyserial>=3.5
requests>=2.31
